requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
physics-tenpy>=0.10.0
qutip>=4.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.20.0
ipywidgets>=8.0.0
tqdm>=4.65.0
numba>=0.57.0
h5py>=3.8.0
pytest>=7.3.0
pytest-cov>=4.0.0 